r551 cleaned or tagged some potentially unsafe or confusing code using -fsanitize=address,integer,undefined added some extra debugging advise for mac to the developer notes

+++ kernel/Utilities/BasisFunctions1DH1ConformingLine.h
@@ -42,7 +42,7 @@ namespace Utilities
     {
     public:
         BasisFunction1DVertexLine(std::size_t node)
-                : nodePosition_(2 * node - 1)
+                : nodePosition_(2 * static_cast<int>(node) - 1)
         {
             logger.assert(node < 2, "A line only has 2 nodes");
         }
